Fix NITTA startup in Python (NITTA-side port selection)
This is a fix for a "port is already in use" occasional problem with
dynamic startup of NITTA API server in Python scripts. Before we've
chosen a random port on Python side and passed it to NITTA via --port.
Sometimes it was already occupied by another process. That caused
random crashes.

This could be fixed more easily by startup retries with port reselection
if only we didn't need to support API server startup right after
synthesis. Synthesis can take a long time, hence greater probability for
chosen ports to become occupied and greater time cost of the retry.

The only robust solution was to implement a NITTA-side free random port
selection and its passing over to Python via inter-process communication.
This commit implements it via Python-side NITTA stdout parsing as this is
the simplest way to achieve the goal.